fix integer conversion warning
git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/trunk@20459 154be670-7969-47af-9ab0-c6fb378d2295

+++ src/Utils/SelectFileFilter.h
@@ -51,7 +51,7 @@ inline void CSelectFileFilter::Load(UINT stringId)
         filtermasks.push_back(temp);
     }
     filterspec.reset(filternames.size());
-    for (int i = 0; i < filternames.size(); ++i)
+    for (size_t i = 0; i < filternames.size(); ++i)
     {
         filterspec[i].pszName = filternames[i];
         filterspec[i].pszSpec = filtermasks[i];
